split utils into contour, rotate, separate_lines
