+++ .gitignore
@@ -2,3 +2,4 @@
 __pycache__
 sbb_newspapers_org_image/pylint.log
 models_eynollah*
+output.html
